Email_assistant/requirements.txt
streamlit==1.29.0
google-api-python-client==2.97.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.1.0
pydantic
langchain-groq
trustcall
langgraph
google-auth
python-dotenv
